requirements.txt
python-dotenv
nextcord
openai
faiss-cpu